third_party/WebKit/LayoutTests/fast/dom/shadow/pseudoclass-update-checked-option-expected.txt

When an option element became 'checked' or not-'checked', distribution should happen.

Since an option element does not create a renderer, we cannot check this using reftest.

PASS nodes1.length is 0
PASS nodes2.length is 1
Flipping :checked state
PASS nodes1.length is 1
PASS nodes2.length is 0
PASS successfullyParsed is true

TEST COMPLETE
